website: revise full development environment instructions
Updates the full development environment instructions to make it clear you *will* need both
Docker and Golangci-Lint installed.

Adds the `poetry-plugin-shell` requirement, now that Poetry requires it.

Updates the per-platform development environment requirements to have a Linux-specific section,
and update the MacOS section to include poetry-plugin-shell and golangci-lint

Moves the instructions on what to do before committing to the bottom of the document; its location
was confusing and didn't clarify what steps were to be taken in what order.

Includes the instruction that, for a first-time run, you must run `make migrate` and `make gen` or
the TS-API won't be built, and in turn the WebUI build would otherwise fail.

We still need instructions for Windows.