web: alter openapi-generator template for models to be Typescript-4 compliant
### What

We use OpenAPI 6.5.0, according to our makefile:

```
docker.io/openapitools/openapi-generator-cli:v6.5.0
```

The template for generating Typescript Model files is in "modelGeneric.mustache."

This commit copies OpenAPI's `typescript/modelGeneric.mustache@6.5.0` into our own template folder
and makes a single change:

```
- export function instanceOf{{classname}}(value: object): boolean {
+ export function instanceOf{{classname}}(value: object): value is {{classname}} {
```

### Why

Typescript 4.0 and up introduced dynamic type guards; these simultaneously assert that an object is
of-a-type at runtime *and* reassure the typescript type checker at build time that we are using an
object correctly.  Typescript compilers strip the type information away, leaving the boolean nature
of the function valid for JavaScript deployments.

Doing this eliminates all the places in our codebase where we have to write our own discriminator.
We have several of these:

```
const isPasswordExpiry(v: object): v is PasswordExpiry => instanceOfPasswordExpiry(v)
```

This is both a waste of CPU and a chore to maintain. Better to eliminate the problem at its source.