# What
Update the CSS to better center the cards on the Flow Stages page.  This commit ensures perfect
centering of our Flow dialogs without breaking the footer, and uses a more concrete & intuitive
layout.

# How

The original specified a width of "35rem"; this commit adopts a more lenient approach,  allowing the
card to shrink-to-fit if there isn't 35rem of space horizontally, exploiting `max-width` and `width:
max()` in a way that allows the card to be big enough in most displays, but to shrink to 95% of the
viewport width when less that 35rem of horizontal space is available.  95% is an arbitrary width;
it gives us just a *little* margin around the card, to retain it's card-ness, without interfering
materially in the transaction.

This specifies that we want this object, which is set by its parent to be the viewport height minus
a uniform margin top and bottom, to be a grid of 32 rows and 1 column.

```css
.ak-login-container {
   max-width: 35rem;
   width: 35rem;
   width: max(35rem, 0.95vw);
   display: grid;
   grid-template-rows: repeat(32, 1fr);
   grid-template-columns: 100%;
}
```

On the card's parent, we tell it to "take up all the room," and then center the card within that
room.  This creates a single grid cell that occupies the entire grid but with perfect centering.

``` css
.ak-login-container > .pf-c-login__main {
    align-self: center;
    grid-row: 1 / -1;
    grid-column: 1 / -1;
}
```

Finally, on the footer, we give it exactly one, bottom row (from the second-to-last to last grid
lines).  Because the card is placed *within* the cell, it creates a new z-index stacking order; in
the unlikely event someone is using a display that's less than ~600px in height, the footer
disappears behind the card.  The card remains interactive even in as little as 320px of height,
although at that point stages like "Change Password" will require a little scrolling to see the
whole card.

``` css
.ak-login-container > .pf-c-login__footer {
    grid-row: -2 / -1;
    grid-column: 1 / -1;
}
```

# Testing

Run and observe.  Use the 📱💻 symbol in DevTools to simulate a variety of resolutions and see it
fit them all.  Run the E2E `good login` tests to assure yourself nothing is broken as a result.

# Observation

Despite the tools telling me that it is perfectly centered, it ends up looking a lower on the page
than my eyeballs think it should.