web/admin: fix language in certificate import modal
There are two buttons on the CertificateKeyPairList: "Create" and "Generate".  This language is confusing, as it is unclear what the difference between those two options could be.  The
second generates and stores a new key pair; the first only creates an entry in our database for an existing pair generated externally. After a short conversation, changing it to
"Import" seemed like a safe option.

\# What

\# Why

\# How

\# Designs

\# Test Steps

\# Other Notes