stages/authenticator_validate: improve autofocus for OTP input field
The same strategy is already implemented by
web/src/flow/components/ak-flow-password-input.ts

Periodically attempt to focus the input field until it is focused. This
is some-what of a crude way to get autofocus, but in most cases the
`autofocus` attribute isn't enough, due to timing within shadow doms and
such.