core/core: limit User search field to pure text
This commit limits the `core/api/users.py:UserViewSet.search_field` entries to a list of fields that, logically, an admin would want to search on when looking for a user. The search is now limited to “username”, “name” (the display
name), and “email”.

[The documentation for DRF states](https://www.django-rest-framework.org/api-guide/filtering/), “The search_fields attribute should be a list of names of text type fields on the model, such as CharField or TextField.” A quick look
through Stack Overflow and Reddit reveals that lots of users have struggled with this restriction, but also that DRF can return highly erroneous lists when non-text fields are included in `search_field`.

We encountered this bug while looking for a user in a very large collection. Usernames with spaces were likely to match on a text segment rather than the whole query, and searches were failing to limit their returns to anything
vaguely resembling the original query string.

Searches for non-text data should be done using the `filterset` feature.